Richmedia adapter : rename files (#9)
