+++ test/spec/modules/big-richmediaBidAdapter_spec.js
@@ -1,5 +1,5 @@
 import { expect } from 'chai';
-import { spec } from 'modules/bigRichmediaBidAdapter.js';
+import { spec } from 'modules/big-richmediaBidAdapter.js';
 import { auctionManager } from 'src/auctionManager.js';
 import * as bidderFactory from 'src/adapters/bidderFactory.js';
 import { config } from 'src/config.js';
